requirements.txt
Flask
gdown
numpy
Pillow
scikit-learn
sklearn
tensorflow
gunicorn
